+++ docs/examples/configure-launcher/model.py
@@ -1,0 +1,14 @@
+"""Dummy Model."""
+
+import logging
+
+
+LOGGER = logging.getLogger(__name__)
+
+
+class Model:
+    def __init__(self, learning_rate: float):
+        self.learning_rate = learning_rate
+
+    def train(self):
+        LOGGER.info(f"Training model with learning_rate {self.learning_rate}")