+++ docs/usage-reference/parser/parser_evaluate_import.py
@@ -1,0 +1,10 @@
+"""EvaluateParser import example."""
+
+import fromconfig
+
+
+if __name__ == "__main__":
+    config = {"_attr_": "str", "_eval_": "import"}
+    parser = fromconfig.parser.EvaluateParser()
+    parsed = parser(config)
+    assert fromconfig.fromconfig(parsed) is str