+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org).
 
 
+## [0.3.1] - 2021-04-21
+
+### Added
+- `DryLauncher` to dry-run config.
+- Skip launcher instantiation if config is `None`
+
+### Changed
+### Deprecated
+### Removed
+### Fixed
+- Wrong order when merging configs in `cli.main`
+
+### Security
+
+
 ## [0.3.0] - 2021-04-20
 
 ### Added